Mmap cleanup. Make sure everything is split sufficiently and is still maintainable.
