--- sa-compression/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "sa-compression"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bitarray = { path = "../bitarray" }
-sa-index = { path = "../sa-index" }
