+++ sa-index/src/array/compressed.rs
@@ -4,84 +4,50 @@ use std::{
 };
 
 use bitarray::{Binary, BitArray, data_to_writer};
-use sa_index::SuffixArray;
+
+use crate::SuffixArray;
 
 /// Writes the compressed suffix array to a writer.
-///
-/// # Arguments
-///
-/// * `sa` - The suffix array to be compressed.
-/// * `sparseness_factor` - The sparseness factor used for compression.
-/// * `bits_per_value` - The number of bits used to represent each value in the compressed array.
-/// * `writer` - The writer to which the compressed array will be written.
-///
-/// # Errors
-///
-/// Returns an error if writing to the writer fails.
 pub fn dump_compressed_suffix_array(
     sa: Vec<i64>,
     sparseness_factor: u8,
     bits_per_value: usize,
     writer: &mut impl Write
 ) -> Result<(), Box<dyn Error>> {
-    // Write the flags to the writer
-    // 00000001 indicates that the suffix array is compressed
-    writer
-        .write(&[bits_per_value as u8])
-        .map_err(|_| "Could not write the required bits to the writer")?;
-
-    // Write the sparseness factor to the writer
-    writer
-        .write(&[sparseness_factor])
-        .map_err(|_| "Could not write the sparseness factor to the writer")?;
-
-    // Write the size of the suffix array to the writer
-    writer
-        .write(&(sa.len() as u64).to_le_bytes())
-        .map_err(|_| "Could not write the size of the suffix array to the writer")?;
-
-    // Compress the suffix array and write it to the writer
-    data_to_writer(sa, bits_per_value, 8 * 1024, writer)
-        .map_err(|_| "Could not write the compressed suffix array to the writer")?;
-
+    writer.write(&[bits_per_value as u8]).map_err(|_| "Could not write the required bits to the writer")?;
+    writer.write(&[sparseness_factor]).map_err(|_| "Could not write the sparseness factor to the writer")?;
+    writer.write(&(sa.len() as u64).to_le_bytes()).map_err(|_| "Could not write the size of the suffix array to the writer")?;
+    data_to_writer(sa, bits_per_value, 8 * 1024, writer).map_err(|_| "Could not write the compressed suffix array to the writer")?;
     Ok(())
 }
 
-/// Load the compressed suffix array from a reader.
-///
-/// # Arguments
-///
-/// * `reader` - The reader from which the compressed array will be read.
-/// * `bits_per_value` - The number of bits used to represent each value in the compressed array.
-///
-/// # Errors
-///
-/// Returns an error if reading from the reader fails.
+/// Load the compressed suffix array, reading the sample_rate + size header first.
+/// Returns a full `SuffixArray::Compressed`.
 pub fn load_compressed_suffix_array(
     reader: &mut impl BufRead,
     bits_per_value: usize
 ) -> Result<SuffixArray, Box<dyn Error>> {
-    // Read the sample rate from the binary file (1 byte)
     let mut sample_rate_buffer = [0_u8; 1];
-    reader
-        .read_exact(&mut sample_rate_buffer)
-        .map_err(|_| "Could not read the sample rate from the binary file")?;
+    reader.read_exact(&mut sample_rate_buffer).map_err(|_| "Could not read the sample rate from the binary file")?;
     let sample_rate = sample_rate_buffer[0];
 
-    // Read the size of the suffix array from the binary file (8 bytes)
     let mut size_buffer = [0_u8; 8];
-    reader
-        .read_exact(&mut size_buffer)
-        .map_err(|_| "Could not read the size of the suffix array from the binary file")?;
+    reader.read_exact(&mut size_buffer).map_err(|_| "Could not read the size of the suffix array from the binary file")?;
     let size = u64::from_le_bytes(size_buffer) as usize;
 
-    // Read the compressed suffix array from the binary file
-    let mut compressed_suffix_array = BitArray::with_capacity(size, bits_per_value);
-    compressed_suffix_array
-        .read_binary(reader)
-        .map_err(|_| "Could not read the compressed suffix array from the binary file")?;
+    let sa = load_compressed(reader, bits_per_value, size)?;
+    Ok(SuffixArray::Compressed(sa, sample_rate))
+}
 
-    Ok(SuffixArray::Compressed(compressed_suffix_array, sample_rate))
+/// Inner helper: load the compressed BitArray body (no header).
+pub(super) fn load_compressed(
+    reader: &mut impl BufRead,
+    bits_per_value: usize,
+    size: usize
+) -> Result<BitArray, Box<dyn Error>> {
+    let mut compressed_suffix_array = BitArray::with_capacity(size, bits_per_value);
+    compressed_suffix_array.read_binary(reader).map_err(|_| "Could not read the compressed suffix array from the binary file")?;
+    Ok(compressed_suffix_array)
 }
 
 #[cfg(test)]
@@ -91,7 +57,6 @@ mod tests {
     use super::*;
 
     pub struct FailingWriter {
-        /// The number of times the write function can be called before it fails.
         pub valid_write_count: usize
     }
 
@@ -100,7 +65,6 @@ mod tests {
             if self.valid_write_count == 0 {
                 return Err(std::io::Error::other("Write failed"));
             }
-
             self.valid_write_count -= 1;
             Ok(1)
         }
@@ -111,7 +75,6 @@ mod tests {
     }
 
     pub struct FailingReader {
-        /// The number of times the read function can be called before it fails.
         pub valid_read_count: usize
     }
 
@@ -120,7 +83,6 @@ mod tests {
             if self.valid_read_count == 0 {
                 return Err(std::io::Error::other("Read failed"));
             }
-
             self.valid_read_count -= 1;
             Ok(buf.len())
         }
@@ -137,15 +99,11 @@ mod tests {
     #[test]
     fn test_dump_compressed_suffix_array() {
         let sa = vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
-
         let mut writer = vec![];
         dump_compressed_suffix_array(sa, 1, 8, &mut writer).unwrap();
 
         assert_eq!(writer, vec![
-            // bits per value
-            8, // sparseness factor
-            1, // size of the suffix array
-            10, 0, 0, 0, 0, 0, 0, 0, // compressed suffix array
+            8, 1, 10, 0, 0, 0, 0, 0, 0, 0,
             8, 7, 6, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 10, 9
         ]);
     }
@@ -154,7 +112,6 @@ mod tests {
     #[should_panic(expected = "Could not write the required bits to the writer")]
     fn test_dump_compressed_suffix_array_fail_required_bits() {
         let mut writer = FailingWriter { valid_write_count: 0 };
-
         dump_compressed_suffix_array(vec![], 1, 8, &mut writer).unwrap();
     }
 
@@ -162,7 +119,6 @@ mod tests {
     #[should_panic(expected = "Could not write the sparseness factor to the writer")]
     fn test_dump_compressed_suffix_array_fail_sparseness_factor() {
         let mut writer = FailingWriter { valid_write_count: 1 };
-
         dump_compressed_suffix_array(vec![], 1, 8, &mut writer).unwrap();
     }
 
@@ -170,7 +126,6 @@ mod tests {
     #[should_panic(expected = "Could not write the size of the suffix array to the writer")]
     fn test_dump_compressed_suffix_array_fail_size() {
         let mut writer = FailingWriter { valid_write_count: 2 };
-
         dump_compressed_suffix_array(vec![], 1, 8, &mut writer).unwrap();
     }
 
@@ -178,15 +133,16 @@ mod tests {
     #[should_panic(expected = "Could not write the compressed suffix array to the writer")]
     fn test_dump_compressed_suffix_array_fail_compressed_suffix_array() {
         let mut writer = FailingWriter { valid_write_count: 3 };
-
         dump_compressed_suffix_array(vec![1], 1, 8, &mut writer).unwrap();
     }
 
     #[test]
     fn test_load_compressed_suffix_array() {
-        let data = [1, // size of the suffix array
-            10, 0, 0, 0, 0, 0, 0, 0, // compressed suffix array
-            8, 7, 6, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 10, 9];
+        let data = [
+            1,
+            10, 0, 0, 0, 0, 0, 0, 0,
+            8, 7, 6, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 10, 9
+        ];
 
         let mut reader = std::io::BufReader::new(&data[..]);
         let compressed_suffix_array = load_compressed_suffix_array(&mut reader, 8).unwrap();
@@ -201,7 +157,6 @@ mod tests {
     #[should_panic(expected = "Could not read the sample rate from the binary file")]
     fn test_load_compressed_suffix_array_fail_sample_rate() {
         let mut reader = FailingReader { valid_read_count: 0 };
-
         load_compressed_suffix_array(&mut reader, 8).unwrap();
     }
 
@@ -209,7 +164,6 @@ mod tests {
     #[should_panic(expected = "Could not read the size of the suffix array from the binary file")]
     fn test_load_compressed_suffix_array_fail_size() {
         let mut reader = FailingReader { valid_read_count: 1 };
-
         load_compressed_suffix_array(&mut reader, 8).unwrap();
     }
 
@@ -217,7 +171,6 @@ mod tests {
     #[should_panic(expected = "Could not read the compressed suffix array from the binary file")]
     fn test_load_compressed_suffix_array_fail_compressed_suffix_array() {
         let mut reader = FailingReader { valid_read_count: 2 };
-
         load_compressed_suffix_array(&mut reader, 8).unwrap();
     }
