+++ sa-index/src/suffix_to_protein_index/dense.rs
@@ -1,0 +1,181 @@
+use std::io::{Read, Write};
+use std::error::Error;
+
+use memmap2::Mmap;
+use sa_mappings::proteins::{SEPARATION_CHARACTER, TERMINATION_CHARACTER};
+use text_compression::ProteinText;
+
+use crate::Nullable;
+use super::SuffixToProteinIndex;
+
+/// Mapping that uses O(n) memory with n the size of the input text, but retrieval of the protein is
+/// in O(1)
+#[derive(Debug, PartialEq)]
+pub struct DenseSuffixToProtein {
+    // UniProtKB does not have more that u32::MAX proteins, so a larger type is not needed
+    mapping: Vec<u32>
+}
+
+/// Mapping backed by a memory-mapped Dense binary file.
+/// Format: [1 byte type=0x00] [8 bytes count (u64 LE)] [count × 4 bytes (u32 LE)]
+pub struct MmapDenseSuffixToProtein {
+    pub(super) mmap: Mmap,
+    pub(super) data_offset: usize, // 9 = 1 (type) + 8 (count)
+}
+
+impl SuffixToProteinIndex for DenseSuffixToProtein {
+    fn suffix_to_protein(&self, suffix: i64) -> u32 {
+        self.mapping[suffix as usize]
+    }
+}
+
+impl SuffixToProteinIndex for MmapDenseSuffixToProtein {
+    fn suffix_to_protein(&self, suffix: i64) -> u32 {
+        let off = self.data_offset + suffix as usize * 4;
+        u32::from_le_bytes(self.mmap[off..off + 4].try_into().unwrap())
+    }
+}
+
+impl DenseSuffixToProtein {
+    /// Creates a new DenseSuffixToProtein mapping
+    ///
+    /// # Arguments
+    /// * `text` - The text over which we want to create the mapping
+    ///
+    /// # Returns
+    ///
+    /// Returns a new DenseSuffixToProtein build over the provided text
+    pub fn new(text: &ProteinText) -> Self {
+        let mut current_protein_index: u32 = 0;
+        let mut suffix_index_to_protein: Vec<u32> = vec![];
+        for char in text.iter() {
+            if char == SEPARATION_CHARACTER || char == TERMINATION_CHARACTER {
+                current_protein_index += 1;
+                suffix_index_to_protein.push(u32::NULL);
+            } else {
+                assert_ne!(current_protein_index, u32::NULL);
+                suffix_index_to_protein.push(current_protein_index);
+            }
+        }
+        suffix_index_to_protein.shrink_to_fit();
+        DenseSuffixToProtein { mapping: suffix_index_to_protein }
+    }
+}
+
+pub(super) fn write_dense_mapping<W: Write>(mapping: &DenseSuffixToProtein, writer: &mut W) -> Result<(), Box<dyn Error>> {
+    let count = mapping.mapping.len() as u64;
+    writer.write_all(&count.to_le_bytes())?;
+    for &val in &mapping.mapping {
+        writer.write_all(&val.to_le_bytes())?;
+    }
+    Ok(())
+}
+
+pub(super) fn read_dense_mapping<R: Read>(reader: &mut R) -> Result<DenseSuffixToProtein, Box<dyn Error>> {
+    let mut buf8 = [0u8; 8];
+    reader.read_exact(&mut buf8)?;
+    let count = u64::from_le_bytes(buf8) as usize;
+    let mut mapping = Vec::with_capacity(count);
+    for _ in 0..count {
+        let mut buf4 = [0u8; 4];
+        reader.read_exact(&mut buf4)?;
+        mapping.push(u32::from_le_bytes(buf4));
+    }
+    Ok(DenseSuffixToProtein { mapping })
+}
+
+#[cfg(test)]
+mod tests {
+    use std::io::Cursor;
+    use std::io::Write as IoWrite;
+
+    use sa_mappings::proteins::{SEPARATION_CHARACTER, TERMINATION_CHARACTER};
+    use text_compression::ProteinText;
+
+    use crate::{Nullable, ReadBinaryMmap};
+    use crate::suffix_to_protein_index::{SuffixToProteinIndex, SuffixToProteinMapping};
+    use super::{DenseSuffixToProtein, write_dense_mapping, read_dense_mapping};
+
+    fn build_text() -> ProteinText {
+        let mut text = ["ACG", "CG", "AAA"].join(&format!("{}", SEPARATION_CHARACTER as char));
+        text.push(TERMINATION_CHARACTER as char);
+        ProteinText::from_string(&text)
+    }
+
+    fn write_to_tempfile(buf: &[u8]) -> tempfile::NamedTempFile {
+        let mut tmp = tempfile::NamedTempFile::new().unwrap();
+        tmp.write_all(buf).unwrap();
+        tmp.flush().unwrap();
+        tmp
+    }
+
+    #[test]
+    fn test_dense_build() {
+        let u8_text = &build_text();
+        let index = DenseSuffixToProtein::new(u8_text);
+        let expected = DenseSuffixToProtein {
+            mapping: vec![0, 0, 0, u32::NULL, 1, 1, u32::NULL, 2, 2, 2, u32::NULL]
+        };
+        assert_eq!(index, expected);
+    }
+
+    #[test]
+    fn test_search_dense() {
+        let u8_text = &build_text();
+        let index = DenseSuffixToProtein::new(u8_text);
+        assert_eq!(index.suffix_to_protein(5), 1);
+        assert_eq!(index.suffix_to_protein(7), 2);
+        // suffix that starts with SEPARATION_CHARACTER
+        assert_eq!(index.suffix_to_protein(3), u32::NULL);
+        // suffix that starts with TERMINATION_CHARACTER
+        assert_eq!(index.suffix_to_protein(10), u32::NULL);
+    }
+
+    #[test]
+    fn test_dense_roundtrip() {
+        let text = build_text();
+        let original = DenseSuffixToProtein::new(&text);
+        let mut buf = Vec::new();
+        write_dense_mapping(&original, &mut buf).unwrap();
+        let mut cursor = Cursor::new(buf);
+        let restored = read_dense_mapping(&mut cursor).unwrap();
+        assert_eq!(original, restored);
+    }
+
+    #[test]
+    fn test_mmap_dense_roundtrip() {
+        let text = build_text();
+        let original = DenseSuffixToProtein::new(&text);
+
+        let mut buf = Vec::new();
+        buf.push(0u8); // type byte
+        write_dense_mapping(&original, &mut buf).unwrap();
+
+        let tmp = write_to_tempfile(&buf);
+        let loaded = SuffixToProteinMapping::read_binary_mmap(tmp.path()).unwrap().0;
+
+        for i in 0..text.len() as i64 {
+            assert_eq!(
+                original.suffix_to_protein(i),
+                loaded.suffix_to_protein(i),
+                "mismatch at suffix {}",
+                i
+            );
+        }
+    }
+
+    #[test]
+    fn test_search_mmap_dense() {
+        let text = build_text();
+        let mut buf = Vec::new();
+        buf.push(0u8);
+        write_dense_mapping(&DenseSuffixToProtein::new(&text), &mut buf).unwrap();
+        let tmp = write_to_tempfile(&buf);
+        let index = SuffixToProteinMapping::read_binary_mmap(tmp.path()).unwrap().0;
+
+        assert_eq!(index.suffix_to_protein(5), 1);
+        assert_eq!(index.suffix_to_protein(7), 2);
+        assert_eq!(index.suffix_to_protein(3), u32::NULL); // SEPARATION_CHARACTER
+        assert_eq!(index.suffix_to_protein(10), u32::NULL); // TERMINATION_CHARACTER
+    }
+}
