+++ sa-server/Cargo.toml
@@ -12,4 +12,3 @@ tokio = { version = "1.36.0", features = ["rt", "rt-multi-thread", "macros"] }
 sa-index = { path = "../sa-index" }
 clap = { version = "4.5.1", features = ["derive"] }
 sa-mappings = { path = "../sa-mappings" }
-sa-compression = { path = "../sa-compression" }
