+++ sa-server/src/lib.rs
@@ -1,45 +1,33 @@
 use std::error::Error;
 use std::fs::File;
-use std::io::{BufReader, Read};
-use sa_compression::load_compressed_suffix_array;
-use sa_index::binary::{load_suffix_array, load_suffix_array_mmap};
-use sa_index::suffix_to_protein_index::{SuffixToProteinIndex, load_mapping, load_mapping_mmap};
-use sa_index::SuffixArray;
+use std::io::BufReader;
+use sa_index::{ReadBinary, ReadBinaryMmap, SuffixArray};
+use sa_index::suffix_to_protein_index::SuffixToProteinMapping;
 use sa_mappings::proteins::Proteins;
 
 pub fn load_suffix_array_file(file: &str, use_mmap: bool) -> Result<SuffixArray, Box<dyn Error>> {
     if use_mmap {
-        return load_suffix_array_mmap(std::path::Path::new(file));
+        return SuffixArray::read_binary_mmap(std::path::Path::new(file));
     }
-
-    let mut sa_file = File::open(file)?;
-    let mut reader = BufReader::new(&mut sa_file);
-
-    let mut bits_per_value_buffer = [0_u8; 1];
-    reader
-        .read_exact(&mut bits_per_value_buffer)
-        .map_err(|_| "Could not read the flags from the binary file")?;
-    let bits_per_value = bits_per_value_buffer[0];
-
-    if bits_per_value == 64 {
-        load_suffix_array(&mut reader)
-    } else {
-        load_compressed_suffix_array(&mut reader, bits_per_value as usize)
-    }
+    let f = File::open(file)?;
+    let mut reader = BufReader::new(f);
+    SuffixArray::read_binary(&mut reader)
 }
 
 pub fn load_proteins_file(file: &str, use_mmap: bool) -> Result<Proteins, Box<dyn Error>> {
     if use_mmap {
-        return Proteins::load_from_binary_mmap(file);
-    }
-    Proteins::load_from_binary(file)
-}
-
-pub fn load_mapping_file(file: &str, use_mmap: bool) -> Result<Box<dyn SuffixToProteinIndex>, Box<dyn Error>> {
-    if use_mmap {
-        return load_mapping_mmap(std::path::Path::new(file));
+        return Proteins::read_binary_mmap(std::path::Path::new(file));
     }
     let f = File::open(file)?;
     let mut reader = BufReader::new(f);
-    load_mapping(&mut reader)
+    Proteins::read_binary(&mut reader)
+}
+
+pub fn load_mapping_file(file: &str, use_mmap: bool) -> Result<SuffixToProteinMapping, Box<dyn Error>> {
+    if use_mmap {
+        return SuffixToProteinMapping::read_binary_mmap(std::path::Path::new(file));
+    }
+    let f = File::open(file)?;
+    let mut reader = BufReader::new(f);
+    SuffixToProteinMapping::read_binary(&mut reader)
 }
